+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -17,9 +17,9 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
+import roomescape.domain.repository.ReservationRepository;
 import roomescape.exception.reservation.NotFoundReservationException;
 
 @Repository
