+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -14,7 +14,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 import roomescape.domain.Theme;
-import roomescape.domain.ThemeRepository;
+import roomescape.domain.repository.ThemeRepository;
 import roomescape.exception.theme.NotFoundThemeException;
 
 @Repository
