+++ src/main/java/roomescape/domain/repository/ReservationTimeRepository.java
@@ -1,8 +1,10 @@
-package roomescape.domain;
+package roomescape.domain.repository;
 
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
+
+import roomescape.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
     List<ReservationTime> findAll();
