+++ src/main/java/roomescape/domain/repository/ThemeRepository.java
@@ -1,7 +1,9 @@
-package roomescape.domain;
+package roomescape.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+
+import roomescape.domain.Theme;
 
 public interface ThemeRepository {
     List<Theme> findAll();
