+++ src/main/java/roomescape/service/UserService.java
@@ -3,12 +3,12 @@ package roomescape.service;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import roomescape.domain.UserRepository;
+import roomescape.domain.repository.UserRepository;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
 
-    
+
 }
